Restructure src directory from template
