+++ v0.3-ts/src/App/App.test.tsx
@@ -1,9 +1,12 @@
+
+// @ts-ignore
 import React from 'react';
 import { render } from '@testing-library/react';
 import App from './App';
 
-test('renders learn react link', () => {
+it('renders learn react link', () => {
   const { getByText } = render(<App />);
   const linkElement = getByText(/learn react/i);
+  // @ts-ignore
   expect(linkElement).toBeInTheDocument();
 });
