+++ v0.3-ts/src/App/App.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>
+      Hello World!
+    </div>
+  );
+}
+
+export default App;
