+++ v0.3-ts/src/index.scss
@@ -1,0 +1,5 @@
+
+* {
+  padding: 0;
+  margin: 0;
+}
