+++ v0.3-ts/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import './index.scss';
+import App from './App/App';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
